Acey, Snakes, move Telari and Isen Dun quests into their own directories.
