+++ scripts/env.lua
@@ -3,6 +3,9 @@ package.path = package.path .. ";./scripts/?.lua"
 package.path = package.path .. ";./scripts/common/?.lua"
 package.path = package.path .. ";./scripts/quests/?.lua"
 package.path = package.path .. ";./scripts/quests/carcassia/?.lua"
+package.pack = package.path .. ";./scripts/quests/isen_dun/?.lua"
+package.path = package.path .. ";./scripts/quests/sithrin/?.lua"
+package.path = package.path .. ";./scripts/quests/telari/?.lua"
 package.path = package.path .. ";./scripts/level/?.lua"
 package.path = package.path .. ";./scripts/areas/?.lua"
 package.path = package.path .. ";./scripts/areas/events/?.lua"
