+++ scripts/quests/carcassia/acey.lua
@@ -1,0 +1,1 @@
+clear_and_add_message("ACEY_SPEECH_TEXT_SID")
