+++ scripts/quests/isen_dun/cynwise.lua
