+++ scripts/quests/isen_dun/hrimgar.lua
