+++ scripts/quests/isen_dun/idem.lua
