+++ scripts/quests/isen_dun/kew.lua
