+++ scripts/quests/isen_dun/nama.lua
