+++ scripts/quests/isen_dun/the_blacksmith.lua
