+++ scripts/quests/isen_dun/the_ploughman.lua
