+++ scripts/quests/sithrin/snakes.lua
@@ -1,0 +1,1 @@
+clear_and_add_message("SNAKES_SPEECH_TEXT_SID")
