--- scripts/quests/the_sentry.lua
@@ -1,5 +1,0 @@
-if get_num_deities() == 0 then
-  clear_and_add_message("THE_SENTRY_GODSLAYER_SPEECH_TEXT_SID")
-else
-  clear_and_add_message("THE_SENTRY_SPEECH_TEXT_SID")
-end
