+++ shadowofthewyrmtext_blank.ini
@@ -2578,6 +2578,14 @@ GUILDMASTER_CARO_DISARM_TRAPS_QUEST_TITLE_SID=
 GUILDMASTER_CARO_DISARM_TRAPS_DESCRIPTION_SID=
 GUILDMASTER_CARO_DISARM_TRAPS_QUEST_COMPLETE_SID=
 GUILDMASTER_CARO_DISARM_TRAPS_QUEST_REMINDER_SID=
+ACEY_SHORT_DESCRIPTION_SID=
+ACEY_DESCRIPTION_SID=
+ACEY_TEXT_DETAILS_SID=
+ACEY_SPEECH_TEXT_SID=
+SNAKES_SHORT_DESCRIPTION_SID=
+SNAKES_DESCRIPTION_SID=
+SNAKES_TEXT_DETAILS_SID=
+SNAKES_SPEECH_TEXT_SID=
 CARCASSIAN_GUARD_SHORT_DESCRIPTION_SID=
 CARCASSIAN_GUARD_DESCRIPTION_SID=
 CARCASSIAN_GUARD_TEXT_DETAILS_SID=
